Add template box display file
Adds deck.js script that displays a 3d box that
can be adjusted in size by the user.

This is based on a sandbox file that was being used
to learn about three.js. Copying this code over
rather than forking because this is a new stand-
alone project.

deck.js code is structured organized similarly to
the three.js sdf to mesh example:
https://github.com/mrdoob/three.js/blob/master/examples/webgl_geometry_sdf.html